fix(release-notes)!: remove context variable from environment
BREAKING CHANGE: This change removes the context.history value from the jinja
templating environment. Since the release notes only document the current release
version, it does not make sense to provide the entire history of releases to the
environment for parsing. As it is customary to provide a link to changes between
the last version and the new version, we now provide a few global variables that
identify the key versions that are generally used (like prev_version, etc.).

See the v10 migration notes and the updated documentation to use the new version
variables in your release notes templates.